agentoz-rag/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
llama-index>=0.10.0
llama-index-llms-dashscope
llama-index-embeddings-dashscope
llama-index-vector-stores-tair
tair>=1.3.0
openai>=1.0.0
PyMuPDF>=1.23.0
requests>=2.31.0
